requirements.txt
pandas == 1.3.0
psycopg2-binary == 2.9.1
openpyxl == 3.0.7
Pillow == 8.2.0



